qcacmn: Populate timestamp,duration,lsig_a stats
ppdu stats timestamp,duration,channel etc were not
updated and corrected the same

Also RSSI was incorrectly getting updated for
mcast frames also, fix both of the above issues

+++ dp/inc/cdp_txrx_cmn_struct.h
@@ -1037,7 +1037,7 @@ struct cdp_rx_indication_ppdu {
 	} u;
 	uint32_t lsig_a;
 	uint32_t rssi;
-	uint32_t timestamp;
+	uint64_t timestamp;
 	uint8_t channel;
 };
